+++ janet-http/http-clients/okhttp3/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'java'
-apply plugin: 'provided-base'
 apply plugin: 'maven'
 
 sourceCompatibility = 1.6
@@ -23,6 +22,6 @@ artifacts {
 
 dependencies {
     compile project(':janet-http:http-clients:url-connection')
-    compile 'com.squareup.okhttp:okhttp:2.7.5'
-    compile 'com.squareup.okhttp:okhttp-urlconnection:2.7.5'
+    compile 'com.squareup.okhttp3:okhttp:3.2.0'
+    compile 'com.squareup.okhttp3:okhttp-urlconnection:3.2.0'
 }
