+++ samples/sample-java/build.gradle
@@ -17,6 +17,7 @@ dependencies {
     compile project(":janet-http:http-service")
     compile project(":janet-http:http-service-compiler")
     compile project(":janet-http:http-clients:okhttp")
+    compile project(":janet-http:http-clients:okhttp3")
     compile project(":janet-http:http-clients:android-apache-client")
     compile project(":janet-async:async-service")
     compile project(":janet-async:async-service-compiler")
