+++ settings.gradle
@@ -5,6 +5,7 @@ include ':janet',
         ':janet-http:http-service-compiler',
         ':janet-http:http-clients:android-apache-client',
         ':janet-http:http-clients:okhttp',
+        ':janet-http:http-clients:okhttp3',
         ':janet-http:http-clients:url-connection',
         ':janet-async:async-service',
         ':janet-async:async-service-compiler',
